Relocated data handlers to own folder
